fix(test): extract shared test utilities and fix infrastructure issues
- Extract MOCK_USER and EVENT_DEFAULTS into shared test-utils.ts with
  createMockLoadEvent() factory and createMockCookies() helper
- Update root layout, (app) layout, and login page tests to import
  from shared module instead of duplicating setup
- Fix vi.stubGlobal leak in health.test.ts by adding vi.unstubAllGlobals()
  in afterEach and re-stubbing document in beforeEach
- Remove passWithNoTests: true from vitest config now that the project
  has substantial test coverage

Closes #300
