+++ src/frontend/src/routes/(public)/login/page.server.test.ts
@@ -1,44 +1,9 @@
 import { describe, expect, it, vi } from 'vitest';
 import { isRedirect } from '@sveltejs/kit';
 import { load } from './+page.server';
+import { MOCK_USER, createMockLoadEvent } from '../../../test-utils';
 
 type LoadEvent = Parameters<typeof load>[0];
-
-const MOCK_USER = {
-	id: '00000000-0000-0000-0000-000000000001',
-	username: 'test@example.com',
-	email: 'test@example.com',
-	firstName: 'Test',
-	lastName: 'User',
-	roles: ['User'],
-	permissions: [],
-	emailConfirmed: true
-};
-
-/** Stubs for all `ServerLoadEvent` properties the load function does NOT use. */
-const EVENT_DEFAULTS = {
-	cookies: {
-		get: vi.fn(),
-		getAll: vi.fn(() => []),
-		set: vi.fn(),
-		delete: vi.fn(),
-		serialize: vi.fn()
-	},
-	fetch: vi.fn() as typeof fetch,
-	getClientAddress: () => '127.0.0.1',
-	locals: { user: null, locale: 'en' },
-	params: {},
-	platform: undefined,
-	request: new Request('http://localhost'),
-	route: { id: '/(public)/login' },
-	setHeaders: vi.fn(),
-	isDataRequest: false,
-	isSubRequest: false,
-	isRemoteRequest: false,
-	tracing: { enabled: false, root: {}, current: {} },
-	depends: vi.fn(),
-	untrack: <T>(fn: () => T): T => fn()
-};
 
 /** Builds a complete mock SvelteKit load event for the login page. */
 function mockLoadEvent(
@@ -54,11 +19,11 @@ function mockLoadEvent(
 		url.searchParams.set(key, value);
 	}
 
-	return {
-		...EVENT_DEFAULTS,
+	return createMockLoadEvent({
+		route: { id: '/(public)/login' },
 		parent: vi.fn().mockResolvedValue({ user }),
 		url
-	} as LoadEvent;
+	}) as LoadEvent;
 }
 
 describe('login page server load', () => {
