+++ src/frontend/vite.config.ts
@@ -18,7 +18,6 @@ export default defineConfig({
 		include: ['src/**/*.test.ts'],
 		environment: 'node',
 		setupFiles: ['src/test-setup.ts'],
-		restoreMocks: true,
-		passWithNoTests: true
+		restoreMocks: true
 	}
 });
